+++ src/test/java/io/kyberorg/yalsee/test/ui/home/VisibleStateTest.java
@@ -57,7 +57,7 @@ public class VisibleStateTest extends SelenideTest {
     }
 
     /**
-     *  Tests that area with overall info exists and visible.
+     * Tests that area with overall info exists and visible.
      */
     @Test
     public void overallAreaIsVisible() {
@@ -79,6 +79,14 @@ public class VisibleStateTest extends SelenideTest {
     @Test
     public void qrCodeAreaIsHidden() {
         HomePageObject.QrCodeArea.QR_CODE_AREA.shouldNotBe(visible);
+    }
+
+    /**
+     * Tests that by default area with My Links Note Area is hidden.
+     */
+    @Test
+    public void myLinksNoteAreaIsHidden() {
+        HomePageObject.MyLinksNoteArea.MY_LINKS_NOTE_AREA.shouldNotBe(visible);
     }
 
     /**